Improved dashboard container resizing
Removed height: 100% from map component container that was preventing resizing from feeling natural
Swapped react-resize for re-resize because it allowed resizing along an entire edge - with react-resize, you were limited to a small (configureably sized) rectangle. This also solves the runtime error of not providing a width prop for the resizable box, as re-resize does not require a width to be specified.